+++ examples/with-spark/package.json
@@ -9,10 +9,8 @@
     "e2e:regtest": "tsx src/e2e-regtest.ts",
     "e2e:lightning-regtest": "tsx src/e2e-lightning-regtest.ts",
     "token-transfer": "tsx src/token-transfer.ts",
-    "deposit": "tsx src/deposit-turnkey.ts",
-    "deposit:turnkey": "tsx src/deposit-turnkey.ts",
-    "deposit:static": "tsx src/static-deposit-turnkey.ts",
-    "deposit:static-turnkey": "tsx src/static-deposit-turnkey.ts",
+    "deposit": "tsx src/scripts/deposit-turnkey.ts",
+    "deposit:static": "tsx src/scripts/static-deposit-turnkey.ts",
     "transfer": "tsx src/transfer.ts",
     "claim": "tsx src/claim.ts",
     "lightning:receive": "tsx src/lightning-receive.ts",
