+++ examples/with-spark/src/e2e-lightning-regtest.ts
@@ -22,10 +22,8 @@ import {
   createTurnkeyLightningInvoice,
   turnkeyPayLightningInvoice,
 } from "./turnkeyLightning";
-import {
-  DEFAULT_SPARK_REGTEST_ELECTRS_URL,
-  depositTurnkeyL1ToSpark,
-} from "./turnkeyL1Deposit";
+import { DEFAULT_SPARK_REGTEST_ELECTRS_URL } from "./spark-deposit/common";
+import { depositTurnkeyL1ToSpark } from "./spark-deposit/normal";
 
 type PendingTransfer = {
   id: string;
