+++ examples/with-spark/src/scripts/deposit-turnkey.ts
@@ -23,11 +23,9 @@
  */
 
 import { Turnkey as TurnkeyServerSDK } from "@turnkey/sdk-server";
-import { env, initSparkWallet, requireEnv } from "./init";
-import {
-  DEFAULT_SPARK_REGTEST_ELECTRS_URL,
-  depositTurnkeyL1ToSpark,
-} from "./turnkeyL1Deposit";
+import { env, initSparkWallet, requireEnv } from "../init";
+import { DEFAULT_SPARK_REGTEST_ELECTRS_URL } from "../spark-deposit/common";
+import { depositTurnkeyL1ToSpark } from "../spark-deposit/normal";
 
 function optionalBigIntEnv(name: string): bigint | undefined {
   const value = process.env[name];
